Move all code from `new_etl` to parent directory
