+++ data/src/classes/data_diff.py
@@ -2,7 +2,7 @@ from datetime import datetime
 import os
 import re
 
-from src.new_etl.classes.file_manager import FileManager, FileType, LoadType
+from src.classes.file_manager import FileManager, FileType, LoadType
 
 file_manager = FileManager()
