+++ data/src/classes/featurelayer.py
@@ -18,9 +18,9 @@ from src.config.config import (
     min_tiles_file_size_in_bytes,
     write_production_tiles_file,
 )
-from src.new_etl.classes.bucket_manager import GCSBucketManager
-from src.new_etl.classes.file_manager import FileManager, FileType, LoadType
-from src.new_etl.loaders import load_carto_data, load_esri_data
+from src.classes.bucket_manager import GCSBucketManager
+from src.classes.file_manager import FileManager, FileType, LoadType
+from src.loaders import load_carto_data, load_esri_data
 
 log.basicConfig(level=log_level)
