+++ data/src/data_utils/city_owned_properties.py
