+++ data/src/data_utils/gun_crimes.py
@@ -1,4 +1,4 @@
-from src.new_etl.data_utils.kde import apply_kde_to_primary
+from src.data_utils.kde import apply_kde_to_primary
 
 from ..classes.featurelayer import FeatureLayer
 from ..constants.services import GUNCRIME_SQL_QUERY
