+++ data/src/data_utils/opa_properties.py
@@ -2,7 +2,7 @@ import re
 
 import pandas as pd
 
-from src.new_etl.metadata.metadata_utils import provide_metadata
+from src.metadata.metadata_utils import provide_metadata
 
 from ..classes.featurelayer import FeatureLayer
 from ..constants.services import OPA_PROPERTIES_QUERY
