+++ data/src/data_utils/ppr_properties.py
