+++ data/src/data_utils/priority_level.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from src.new_etl.metadata.metadata_utils import provide_metadata
+from src.metadata.metadata_utils import provide_metadata
 from ..classes.featurelayer import FeatureLayer
