+++ data/src/data_utils/vacant_properties.py
