+++ data/src/test/test_metadata_utils.py
@@ -2,8 +2,8 @@ import unittest
 
 import geopandas as gpd
 
-from src.new_etl.classes.featurelayer import FeatureLayer
-from src.new_etl.metadata.metadata_utils import (
+from src.classes.featurelayer import FeatureLayer
+from src.metadata.metadata_utils import (
     get_column_details,
     get_description_from_docstring,
     get_sections_from_docstring,
