+++ data/src/validation/city_owned_properties.py
