+++ data/src/validation/ppr_properties.py
