+++ data/src/validation/vacant_properties.py
